Pro ASP.Net MVC 5 - SportsStore - Ch 12 - More Changes
Changed the hardcoded password to something harder to hack.
Added a robots.txt file to keep search engines from listing the site.
Updated the mobile UI to use the thumbnail with the description in the alt
attribute rather than show the description, given the tiny screen space on the
iPhone.