Pro ASP.Net MVC 5 Platform - Ch 13 - Initial Admin Controllers
Fixed a configuration issue because I created the project with .Net 4.5 and
then changed it to .Net 4.5.1. I needed to update all the NuGet packages. It
was showing errors in the views, as if Visual Studio had trouble resolving
some classes for Intellisense.

Created an Admin controller and a view that enumerates users. Now need to add
the rest of the CRUD.