Pro ASP.Net MVC 5 Platform - Ch 14 - Added DB Seeding Code
Added simple seeding code that creates an Administrators role and an
administrator account. Since it's simple, it requires that the DB be brand new,
so the previous copy needed to be deleted and then added again.
There was a weird failure when the AppUserManager.Create method was called, the
user never got created and no exception was thrown. I changed it to store the
return value of that call into a variable that would then throw an exception
if the result did not denote success. It worked this time, but I'm not
convinced that is the fix. Definitely a problem with this new Identity 2.0 API.